+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -156,7 +156,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-06-02T09:50:17.688Z",
+    "2026-06-02T13:58:47.485Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {