+++ packages/orq-rc/RELEASES.md
@@ -9669,3 +9669,13 @@ Based on:
 - [typescript v4.10.0-rc.69] packages/orq-rc
 ### Releases
 - [NPM v4.10.0-rc.69] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.69 - packages/orq-rc
+
+## 2026-06-02 14:01:41
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.769.2 (2.892.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.10.0-rc.72] packages/orq-rc
+### Releases
+- [NPM v4.10.0-rc.72] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.72 - packages/orq-rc